# fix: replace CTEs with joins