--- a/test/lib/tables.ts
+++ b/test/lib/tables.ts
@@ -465,12 +465,12 @@ test('primary keys', async () => {
         "name": "t",
         "primary_keys": [
           {
-            "name": "c",
+            "name": "cc",
             "schema": "public",
             "table_name": "t",
           },
           {
-            "name": "cc",
+            "name": "c",
             "schema": "public",
             "table_name": "t",
           },